[Bronze III] Title: 알고리즘 수업 - 알고리즘의 수행 시간 3, Time: 208 ms, Memory: 17676 KB -BaekjoonHub

+++ 백준/Bronze/24264. 알고리즘 수업 － 알고리즘의 수행 시간 3/알고리즘 수업 － 알고리즘의 수행 시간 3.java
@@ -4,7 +4,7 @@ class Main {
 
     public static void main(String[] args) {
         Scanner in = new Scanner(System.in);
-        long n = in.nextInt();
+        long n = in.nextLong();
 
         solution(n);
     }
